+++ core/messages/de/luya.php
@@ -7,5 +7,8 @@ return [
     '{attribute} must include at least one letter.' => '{attribute} muss mindestens einen Buchstaben enthalten.',
     '{attribute} must include at least one uppercase letter.' => '{attribute} muss mindestens einen Grossbuchstaben enthalten.',
     '{attribute} must include at least one lowercase letter.' => '{attribute} muss mindestens einen Kleinbuchstaben enthalten.',
-    '{attribute} must be a float or numeric value.' => '{attribute} muss ein Numerischer Wert sein.'
+    '{attribute} must be a float or numeric value.' => '{attribute} muss ein Numerischer Wert sein.',
+    'Invalid phone number.' => 'Ungültige Telefonnummer.',
+    'The phone number does not match the required type {name}.' => 'Die Telefonnummer entspricht nicht dem erwarteten Typen {name}.',
+    'Invalid phone number, ensure it starts with the correct country code.' => 'Ungültige Telefonnummer; stellen Sie sicher, dass sie mit dem korrekten Ländercode beginnt.',
 ];